Now icon-config.ts only exports module functions.
